backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "agno>=1.7.11",
    "fastapi[standard]>=0.116.1",
    "google-api-python-client>=2.179.0",
    "google-auth>=2.40.3",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "groq>=0.31.0",
    "packaging>=25.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "uvicorn>=0.35.0",
]
